Amy Irving A few notes before you open this
I have pictures in this version. Theyre in absolute path
so to make them work, wdit the path's ode on the
welcomepanel(), Inprocesspanel(), and the FinsihedPanel(), to
see a rad GUI.

If you want to get to the finished panel, but dont see a button,
uncomment the "temp button" in the main GUI.

Sorry about the "inprocess()" panels confusingness.
Its uncommented and i get lost in the mess, Illl organize it
later and get back to you on that.